use GenericModal for new ad form
