+++ frontend/src/components/NewAdModal/modalComponents/AdModalFormAddress.tsx
@@ -72,6 +72,7 @@ export default function AdModalFormAddress({
               field.onChange(value);
               setSelectedSuggestion(value);
             }}
+            noOptionsText="Saissisez une adresse"
             renderInput={(params) => (
               <TextField
                 {...params}
